Transitions: PRIMARY_BOUNCER -> GONE notification scrim
The scrim was incorrectly being left with an alpha of 1f. Make sure
this gets set to 0f to prevent blocking launcher.

Also, when launching animations from keyguard, the bouncer was still
animating away. This seems to have regressed in QPR2. Restore
functionality to hide bouncer immediately so the animation can proceed
from keyguard.

Fixes: 274285526
Test: atest PrimaryBouncerToGoneTransitionViewModelTest
Change-Id: I3d298e037305abdd9479cae2250e3fc945cc7a9a