Revert^2 "Drive SurfaceView visibility on renderthread"
This includes a change to RenderNode which ensures that RenderNode
callbacks are dispatched synchronously in prepareTree(). This is because
the UI thread may be unblocked while replaying drawing commands, and we
don't wait on worker threads for a given frame until after we're done
drawing.

a1353742983854c1104fb2ffdfd03537880fa46d

Bug: 274519506
Bug: 269113414
Test: CtsSurfaceControlTests

Change-Id: Iae5c7d096930ec76fd54d9c41a1a479c86d18870