Fix black screen while animating a closing and changing TF
A TaskFragment could be resizing to a smaller bounds while changing
to be invisible. A black screen was shown during the transition.
So, skip unfreeze surface in that case.

Also union the staring bounds to the whole closing screen bounds,
or the leash will be cropped by the target screen bounds.

Bug: 267044260
Test: steps on the bug
Change-Id: Ia037c5beaca12695fc6c00e38c454cb6dced1b33