Move insetsRoundedCornerFrame into InsetsFrameProvider
The insetsRoundedCornerFrame is used to indicate that the rounded corner
providing to other windows should be moved by the insets provided by the
window, or say, the display should be insets by the window's provided
insets frame when calculating rounded corner related thing.

This is actually using the size of controllable insets of the window, so
moving it to the InsetsFrameProvider to make it bonds to a given insets
source would be a more clean way to describe the issue. Besides, the
original implementation won't actually take effect when the value get
updated at runtime, as the validating and the value set only happens
when adding the window. Modified the update layout params related logic
to make sure the change can happen at runtime, and dispatch
configuration to other windows as needed.

The comments are also rephrased to make it easier to understand.

Bug: 274724339
Test: LetterboxUiControllerTest
Test: SizeCompatTests
Test: Observe task bar behavior on tablet devices.
Change-Id: I9b8aebd340dabefe019148e80bcb3fe480181830