make upgrade version combination resilient
With "==" the version combination was rigid and there could
be edge cases where the exact combination is not matched.
This may result in upgrade steps being not executed and could
impact users. With "<=" we are trying to be more flexible in
terms of version combination and also allowing packages upgrade
even in case of no combination mach.

Fix: 276508167
Test: presubmit/build
Change-Id: If5f489b726c2a82889b16a399ebf63beaa3b5593