Update config according to insets of status bar
Previously, only cutout and navbar are included, so only the change
of decor frame may trigger config change.
But screenWidthDp and screenHeightDp are calculated according to
config (stable) frame, which includes the insets from status bar.

So this change makes the display configuration consistent with the
latest state of the display decor insets.

Fix: 259693044
Test: DisplayPolicyTests#testUpdateDisplayConfigurationByDecor

Change-Id: Idaa7280ed23076b3282c987a249259e0c38b6630