Prepare to log post_duration_millis to statsd
We now compute the milliseconds elapsed and supply the value to the NotificationRecordLogger. But note that nothing is actually logged until we update the proto file.

This CL also removes the mReport field from PostNotificationTracker. After the last-minute code reshuffle in the previous change, it's no longer necessary and we can pass it directly to notifyListenersPostedAndLogLocked().

Bug: 275044361
Test: atest NotificationManagerServiceTest
Change-Id: I23395aaf89d196ca87d1d7325568a46b81c036ea