Use Ctrl+Tab to navigate focus clusters
As part of Android U, the Meta+Tab shortcut now opens the recent
apps. Unfortunately, that blocked that shortcut from arriving to
the app ViewRootImpl, which was currently used to navigate focus
clusters. That means that some elements will not be reachable
using the keyboard, which is a big problem for acessability.

This CL changes the Meta+Tab to Ctrl+Tab, which is not assigned
to any shortcut.

Bug: 278932280
Test: Flashed on device, confirmed that the cluster focus moves
Change-Id: Ib2811bb9d7eeb39f8c184fb4824468d3d1d28c23