Enable users to dismiss keyboard by long-pressing down button
This is a follow up CL to my CL [1], which basically forked

  com.android.systemui.navigationbar.buttons.KeyButtonView

for IMEs.

In the previous implementation, the user could not dismiss the
keyboard by long-pressing the down button.  To address this issue, the
behavior around long-pressing the donw button is changed as follows.

* Before: on long-pressing the down button,
 - KeyEvent.ACTION_DOWN with FLAG_LONG_PRESS is not sent ot the IME.
 - AccessibilityEvent.TYPE_VIEW_LONG_CLICKED is issued.
 - KeyEvent.ACTION_UP is not sent to the IME.

* After: on long-pressing the down button,
 - KeyEvent.ACTION_DOWN with FLAG_LONG_PRESS is not sent ot the IME.
 - AccessibilityEvent.TYPE_VIEW_LONG_CLICKED is not issued.
 - KeyEvent.ACTION_UP is sent to the IME on releasing long-press.

As there is no known usage of long-pressing the down button, not
sending AccessibilityEvent.TYPE_VIEW_LONG_CLICKED should be OK here.

 [1]: I3e7e1f83554444131e2765dc159617bb9e2337c7
      bc36e25fc617e1ca913cbcc8f82b55e5277ed2c7

Bug: 215443343
Fix: 245555619
Test: Manually verified on aosp_bluejay-userdebug
Change-Id: Ib32f8b610197ab57e45b4e5932d5376ddc59c574